requirements.txt
Flask==2.2.5
Flask_RESTful==0.3.9
PyYAML==6.0.1
requests==2.31.0
waitress==2.1.2
Werkzeug==2.3.8
python-dotenv==0.21.0